feat: new option vim.g.nightflySpellErrorColor
The addition of this option is motivated by the behavior of
nightflyUndercurls.

When nightflyUndercurls is enabled, which is the default:
- undercurls are used for spelling and linting errors

However, when nightflyUndercurls is disabled:
- underlining is used for spelling and linting errors, and
- spelling errors are highlighted by setting the foreground color

The new option nightflySpellErrorColor can be used to disable the
highlighting of spelling errors via foreground color. By default this
option is enabled, i.e., underlined spelling errors are also colored.

Note, this new option can NOT be used to enable the highlighting of
spelling errors via foreground color when nightflyUndercurls is enabled.
Since the default for nightflySpellErrorColor is enabled to maintain
current behavior for underlining, doing so would break current behavior
for undercurls.