feat: support normalized Popup Menu colors via new option
Both Vim and Neovim now support bordered Popup Menu (aka for normal
completions).

Historically, without a border, the popup menu needed to be a different
contrasting color to the normal edit window.

But now with a border set the popup menu can be made to use the normal
background color when the new g:nightflyNormalPmenu is set to true
(default is false). This mimics the existing g:nightflyNormalFloat option
(which applies to floating windows).